requirements.txt
# ── Feed Crawler — Dependencies ──

# Core
feedparser==6.0.11
httpx==0.28.1
sqlalchemy==2.0.47
pyyaml==6.0.3
apscheduler==3.11.0
python-dotenv==1.2.1
defusedxml==0.7.1
requests==2.32.5
beautifulsoup4>=4.12
lxml>=5.0

# Database
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20

# Dev / Test
pytest==8.3.5
pytest-asyncio==0.25.3
ruff==0.11.2
